[New Game] : Catch The  Falling Objects
Catch the Falling Objects Game is a browser-based game where players use a basket to catch objects falling from above. Move the basket using your mouse to catch the falling objects. Each caught object earns a point, while missed objects result in a loss of life. The goal is to score as high as possible before running out of lives.